refactor ukdctl start
